Move LayersManayer component to the components folder
