+++ src/components/LayersManager/LayersManager.js
