+++ src/components/LayersManager/LayersSwitcher.js
